Add files via upload
These are a couple of works I did. This includes a Connect Four game in java where we had to build it from scratch, Another assignment in programming where we had to read and write to a file and then read the content of the file and put it inside a Jframe. The other two documents are the last assignment we had to do in business. We had to create a formal report for the company of our choice, and the present parts of it to the class